Fix saved-ref action execution envelopes
